Get Travis-CI working.
Tox defaults to running under {toxinidir}, (locale of tox.ini), so:
The tox test-run with pytest picks up on the code in the base
repository, rather than the code installed into the tox-env.
This confuses pytest (two copies) and it borks.
We get around this by switching dirs and using ``py.test --pyargs <package>``.

We also ensure that we're testing against the master version
of the server-code, by specifying the git-repo as part of the deps.

Also: Install pandoc, define $HOME.